refactor: give the manager its own cohort scope (#225)
매니저는 오퍼레이터와 **다른 조회**로 기수를 정해야 한다. 실계정으로 둘을 나란히
불러 확인했다.

  GET /cohorts                   200 — 그런데 10기·9기·8기·7기 **기관 전체**
  GET /members/me/enrollments    200 — 9기 하나. 담당하는 것만

전자를 쓰면 담당하지 않는 기수가 상단 스위처에 뜨고, 고르는 순간 하위 조회가 전부
담당 반 기준이라 빈 결과로 떨어진다 — 화면은 그걸 「데이터가 없다」로 말하게 된다.

`useManagerCohort()`를 더한다. 순서는 서버가 정해서 준다(`RUNNING → PLANNED →
CLOSED`, 같은 상태 안에서는 시작일 내림차순)이므로 **첫 항목을 그대로 연다** —
오퍼레이터 쪽이 `RUNNING`을 직접 찾는 것과 다르다. 기존 `useCohortId`는 그대로
두고 `?cohort=` 규약만 두 역할이 공유한다.

`_shared/listQuery`를 `lib/`로 올린다. 매니저는 레이어 린트 때문에 operator를
import할 수 없는데, 이미 `app/_async-preview`도 그 파일을 쓰고 있었다 — 세 번째
사용처가 나온 시점이다(D14). 동작은 그대로다.
