+++ src/features/operator/admin/cohorts/CohortsTab.tsx
@@ -12,7 +12,7 @@ import {
   TableCell,
 } from '@/components/ui/Table'
 import { useDebounced } from '@/lib/useDebounced'
-import { listQueryOptions } from '../../_shared/listQuery'
+import { listQueryOptions } from '@/lib/listQuery'
 import { useFindCohorts } from '@/api/academic/useAcademicQueries'
 import { useDeleteCohort, useEndCohort } from '@/api/academic/useAcademicMutations'
 import { isApiError } from '@/api/_contract'
@@ -103,7 +103,7 @@ export default function CohortsTab() {
         size: PAGE_SIZE,
       },
     },
-    /* 조건·페이지를 바꿔도 표를 비우지 않는다 — `_shared/listQuery` 주석 참고 */
+    /* 조건·페이지를 바꿔도 표를 비우지 않는다 — `lib/listQuery` 주석 참고 */
     listQueryOptions,
   )
 
@@ -215,7 +215,7 @@ export default function CohortsTab() {
           </Empty>
         )
       ) : (
-        /* 옛 값을 그리는 동안 그 사실을 숨기지 않는다 — `_shared/listQuery` */
+        /* 옛 값을 그리는 동안 그 사실을 숨기지 않는다 — `lib/listQuery` */
         <StaleBlock stale={page.isFetching && page.data !== undefined} label="기수를 불러오는 중">
           <Table className="table-fixed">
             <TableHeader>
