+++ src/features/operator/admin/curricula/CurriculaTab.tsx
@@ -12,7 +12,7 @@ import {
   TableCell,
 } from '@/components/ui/Table'
 import { useDebounced } from '@/lib/useDebounced'
-import { listQueryOptions } from '../../_shared/listQuery'
+import { listQueryOptions } from '@/lib/listQuery'
 import { useGetCurrentMember } from '@/api/member/useMemberQueries'
 import { useFindOrganizationCurricula } from '@/api/curriculum/useCurriculumQueries'
 import type { findOrganizationCurricula_Query } from '@/api/curriculum/curriculumTypes'
@@ -112,7 +112,7 @@ export default function CurriculaTab() {
         size: PAGE_SIZE,
       },
     },
-    /* 조건·페이지를 바꿔도 표를 비우지 않는다 — `_shared/listQuery` 주석 참고 */
+    /* 조건·페이지를 바꿔도 표를 비우지 않는다 — `lib/listQuery` 주석 참고 */
     { enabled: !!organizationId, ...listQueryOptions },
   )
 
@@ -242,7 +242,7 @@ export default function CurriculaTab() {
           </Empty>
         )
       ) : (
-        /* 옛 값을 그리는 동안 그 사실을 숨기지 않는다 — `_shared/listQuery` */
+        /* 옛 값을 그리는 동안 그 사실을 숨기지 않는다 — `lib/listQuery` */
         <StaleBlock stale={page.isFetching && page.data !== undefined} label="교안을 불러오는 중">
           <Table className="table-fixed">
             <TableHeader>
