+++ src/features/operator/admin/roster/RosterTab.tsx
@@ -13,7 +13,7 @@ import {
   TableCell,
 } from '@/components/ui/Table'
 import { useDebounced } from '@/lib/useDebounced'
-import { listQueryOptions } from '../../_shared/listQuery'
+import { listQueryOptions } from '@/lib/listQuery'
 import { cn } from '@/lib/utils/cn'
 import { useFindClassrooms } from '@/api/academic/useAcademicQueries'
 import { useFindTraineeRoster } from '@/api/member/useMemberQueries'
@@ -143,7 +143,7 @@ export default function RosterTab() {
         size: ROSTER_PAGE_SIZE,
       },
     },
-    /* 조건·페이지를 바꿔도 표를 비우지 않는다 — `_shared/listQuery` 주석 참고 */
+    /* 조건·페이지를 바꿔도 표를 비우지 않는다 — `lib/listQuery` 주석 참고 */
     { enabled: !!cohortId, ...listQueryOptions },
   )
 
@@ -423,7 +423,7 @@ export default function RosterTab() {
           </Empty>
         )
       ) : (
-        /* 옛 값을 그리는 동안 그 사실을 숨기지 않는다 — `_shared/listQuery` */
+        /* 옛 값을 그리는 동안 그 사실을 숨기지 않는다 — `lib/listQuery` */
         <div className={cn('grid gap-4', assignOpen && 'lg:grid-cols-[1fr_300px]')}>
           <div className="min-w-0">
             <StaleBlock
