+++ src/stores/cohortScope.ts
@@ -1,5 +1,18 @@
 import { useSearchParams } from 'react-router'
-import { useFindCohorts } from '@/api/academic/useAcademicQueries'
+import { useFindCohorts, useFindMyEnrollments } from '@/api/academic/useAcademicQueries'
+
+/** 기수 스코프를 쓰는 화면이 헤더 스위처에 그대로 넘기는 모양 */
+export type CohortScope = {
+  cohortId: string | undefined
+  /** 표시명(`9기`). 상단 스위처가 이 값을 그린다 — 이름 때문에 조회를 더 하지 않는다 */
+  cohortName: string | undefined
+  /** 기수가 하나도 없거나 조회가 실패했다 — 화면이 그 자리에 무엇을 그릴지 정한다 */
+  failed: boolean
+  /** 스위처에 넘길 선택지. 아직 안 왔으면 빈 배열이다 */
+  cohorts: { value: string; label: string }[]
+  /** 스위처가 부른다 — 주소를 바꾸면 이 훅을 쓰는 화면이 전부 따라온다 */
+  selectCohort: (cohortId: string) => void
+}
 
 /**
  * 지금 보고 있는 기수 — **화면이 아니라 여기가 갖는다.**
@@ -42,41 +55,20 @@ import { useFindCohorts } from '@/api/academic/useAcademicQueries'
  * `features/operator/admin/_/cohortScope.ts`가 아직 자기 상수를 갖고 있다 — **남의
  * 도메인 파일이라 건드리지 않는다.** 그쪽이 연동될 때 여기로 합친다.
  */
-export function useCohortId(): {
-  cohortId: string | undefined
-  /** 표시명(`9기`). 상단 스위처가 이 값을 그린다 — 이름 때문에 조회를 더 하지 않는다 */
-  cohortName: string | undefined
-  /** 기수가 하나도 없거나 조회가 실패했다 — 화면이 그 자리에 무엇을 그릴지 정한다 */
-  failed: boolean
-  /** 스위처에 넘길 선택지. 아직 안 왔으면 빈 배열이다 */
-  cohorts: { value: string; label: string }[]
-  /** 스위처가 부른다 — 주소를 바꾸면 이 훅을 쓰는 화면이 전부 따라온다 */
-  selectCohort: (cohortId: string) => void
-} {
+export function useCohortId(): CohortScope {
   const { data, isError } = useFindCohorts({ query: { page: 0, size: 50 } })
-  const [params, setParams] = useSearchParams()
+  const { fromUrl, selectCohort } = useCohortParam()
   const list = data?.content ?? []
 
   /*
     주소가 먼저다. 없거나 **목록에 없는 값**이면 진행 중 기수로 떨어진다 —
     남의 조직 기수 id가 주소에 실려 와도 그것으로 조회하지 않는다.
   */
-  const fromUrl = params.get('cohort')
   const picked =
     /* `fromUrl &&` 로 쓰면 빈 문자열이 그대로 흘러 타입이 넓어진다 — 조회를 먼저 한다 */
     (fromUrl ? list.find((c) => c.cohortId === fromUrl) : undefined) ??
     list.find((c) => c.status === 'RUNNING') ??
     list[0]
-
-  const selectCohort = (cohortId: string) => {
-    const next = new URLSearchParams(params)
-    next.set('cohort', cohortId)
-    /*
-      `replace`가 아니다 — 기수를 바꾸는 것은 **다른 것을 보러 가는 것**이라
-      뒤로가기로 되돌아올 수 있어야 한다(검색어 타이핑과 다르다).
-    */
-    setParams(next)
-  }
 
   return {
     cohortId: picked?.cohortId,
@@ -84,5 +76,60 @@ export function useCohortId(): {
     failed: isError || (!!data && list.length === 0),
     cohorts: list.map((c) => ({ value: c.cohortId, label: c.name })),
     selectCohort,
+  }
+}
+
+/**
+ * 매니저가 보고 있는 기수 — **오퍼레이터와 조회가 다르다.**
+ *
+ * `GET /cohorts`(위 `useCohortId`)는 **기관에 개설된 기수 전부**를 준다. 실계정으로
+ * 확인했다 — 매니저 토큰으로도 200이지만 담당하지 않는 `10기`·`8기`·`7기`까지 왔다.
+ * 그걸 스위처에 그리면 **고를 수 있는데 아무것도 안 보이는 기수**가 생긴다: 하위 조회는
+ * 전부 담당 반 기준이라 빈 결과로 떨어지고, 화면은 "데이터가 없다"고 말하게 된다.
+ *
+ * `GET /members/me/enrollments`는 **내가 실제로 속한 기수만** 준다(같은 계정에서 `9기`
+ * 하나). 매니저 스코프의 근거는 이쪽이다.
+ *
+ * **순서를 서버가 정한다**(30차 R9). 종전에는 상태가 안 와서 목록의 첫 기수로 떨어졌고,
+ * 담당 기수가 둘 이상이면 그 순서에 근거가 없었다. 지금은 `RUNNING → PLANNED → CLOSED`,
+ * 같은 상태 안에서는 시작일 내림차순으로 정렬돼 오므로 **첫 항목을 그대로 연다** —
+ * 화면이 다시 고르면 그 규칙이 두 곳에 생기고, 오퍼레이터 쪽(`useCohortId`)이 `RUNNING`을
+ * 직접 찾는 것과 달리 여기서는 서버가 이미 답을 줬다.
+ *
+ * ⚠ **담당 반은 여기서 안 온다.** `enrollments[].classroom`은 매니저에게 늘 `null`이다
+ * (반 배정은 교육생 소속이고, 매니저 배정은 `manager_assignment`라는 다른 원장이다).
+ * 담당 반이 필요한 화면은 `GET /cohorts/{cohortId}/classrooms`를 부른다 — 그 조회가
+ * 매니저에게는 **담당 반만** 돌려준다(실계정: `C반` 하나).
+ */
+export function useManagerCohort(): CohortScope {
+  const { data, isError } = useFindMyEnrollments()
+  const { fromUrl, selectCohort } = useCohortParam()
+  const list = data?.enrollments ?? []
+
+  const picked = (fromUrl ? list.find((e) => e.cohortId === fromUrl) : undefined) ?? list[0]
+
+  return {
+    cohortId: picked?.cohortId,
+    cohortName: picked?.cohortName,
+    failed: isError || (!!data && list.length === 0),
+    cohorts: list.map((e) => ({ value: e.cohortId, label: e.cohortName })),
+    selectCohort,
+  }
+}
+
+/** 고른 기수는 **주소가 갖는다**(`?cohort=`) — 두 역할이 같은 규약을 쓴다 */
+function useCohortParam() {
+  const [params, setParams] = useSearchParams()
+  return {
+    fromUrl: params.get('cohort'),
+    selectCohort: (cohortId: string) => {
+      const next = new URLSearchParams(params)
+      next.set('cohort', cohortId)
+      /*
+        `replace`가 아니다 — 기수를 바꾸는 것은 **다른 것을 보러 가는 것**이라
+        뒤로가기로 되돌아올 수 있어야 한다(검색어 타이핑과 다르다).
+      */
+      setParams(next)
+    },
   }
 }
